+++ app/src/main/java/mashup/kr/mapc/ui/main/ViewPagerAdapter.java
@@ -1,8 +1,13 @@
-package mashup.kr.mapc;
+package mashup.kr.mapc.ui.main;
 
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentPagerAdapter;
+
+import mashup.kr.mapc.ui.main.home.HomeFragment;
+import mashup.kr.mapc.ui.main.magazine.MagazineFragment;
+import mashup.kr.mapc.ui.main.mypage.MypageFragment;
+import mashup.kr.mapc.ui.main.qnp.QnpFragment;
 
 public class ViewPagerAdapter extends FragmentPagerAdapter{
