+++ app/src/main/java/mashup/kr/mapc/ui/main/magazine/MagazineFragment.java
@@ -1,4 +1,4 @@
-package mashup.kr.mapc;
+package mashup.kr.mapc.ui.main.magazine;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
@@ -7,6 +7,8 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import mashup.kr.mapc.R;
 
 public class MagazineFragment extends Fragment{
